Dev/rules engine api router

+++ server/api/controllers/FlowController.js
@@ -124,6 +124,7 @@ module.exports = {
     .then(function(value) {
       var externalID = value.externalID
       flowType = value.flowType
+      published = value.published
       return Q.fcall(function(){
         return {
           externalID: externalID,
@@ -132,8 +133,13 @@ module.exports = {
       })
     })
     .then(function(result){
-      if (flowType == 'genericRule')
+      if (flowType == 'genericRule' && published == true) {
         return RulesEngineService.delete(result)
+      }
+      else
+        return Q.fcall(function (){
+          return {res: {statusCode: '200'}}
+        })
     })
     .then(function(result) {
       if (result.res.statusCode == '200') {
